change navlogo on scroll, add desktopnav container
