+++ src/components/Layout.js
@@ -4,8 +4,7 @@ import { createGlobalStyle } from 'styled-components'
 import { StaticQuery, graphql, withPrefix } from 'gatsby'
 import reset from 'styled-reset'
 
-import Navbar from 'components/Navbar'
-import NavLogo from 'components/NavLogo'
+import DesktopNav from 'components/DesktopNav'
 import Footer from 'components/Footer'
 import MobileHeader from 'components/MobileHeader'
 
@@ -36,8 +35,7 @@ const GlobalStyles = createGlobalStyle`
 export const LayoutTemplate = ({ children, location }) => (
   <>
     <GlobalStyles />
-    <NavLogo location={location} />
-    <Navbar location={location} />
+    <DesktopNav location={location} />
     <MobileHeader location={location} />
     <div>{children}</div>
     <Footer />
